Move create_refund_builder to flow.rs
Move create_refund_builder documentation to flow.rs
