+++ lightning/src/ln/outbound_payment.rs
@@ -434,7 +434,7 @@ impl Display for PaymentAttempts {
 /// [`PendingOutboundPayment::AwaitingOffer`] should be considered stale and candidate for removal
 /// in [`OutboundPayments::remove_stale_payments`].
 #[derive(Clone, Copy)]
-pub(crate) enum StaleExpiration {
+pub enum StaleExpiration {
 	/// Number of times [`OutboundPayments::remove_stale_payments`] is called.
 	TimerTicks(u64),
 	/// Duration since the Unix epoch.
