+++ lightning/src/offers/refund.rs
@@ -77,10 +77,10 @@
 //! # Note
 //!
 //! If constructing a [`Refund`] for use with a [`ChannelManager`], use
-//! [`ChannelManager::create_refund_builder`] instead of [`RefundBuilder::new`].
+//! [`OffersMessageFlow::create_refund_builder`] instead of [`RefundBuilder::new`].
 //!
 //! [`ChannelManager`]: crate::ln::channelmanager::ChannelManager
-//! [`ChannelManager::create_refund_builder`]: crate::ln::channelmanager::ChannelManager::create_refund_builder
+//! [`OffersMessageFlow::create_refund_builder`]: crate::offers::flow::OffersMessageFlow::create_refund_builder
 
 use bitcoin::constants::ChainHash;
 use bitcoin::network::Network;
@@ -159,10 +159,10 @@ macro_rules! refund_explicit_metadata_builder_methods { () => {
 	/// # Note
 	///
 	/// If constructing a [`Refund`] for use with a [`ChannelManager`], use
-	/// [`ChannelManager::create_refund_builder`] instead of [`RefundBuilder::new`].
+	/// [`OffersMessageFlow::create_refund_builder`] instead of [`RefundBuilder::new`].
 	///
 	/// [`ChannelManager`]: crate::ln::channelmanager::ChannelManager
-	/// [`ChannelManager::create_refund_builder`]: crate::ln::channelmanager::ChannelManager::create_refund_builder
+	/// [`OffersMessageFlow::create_refund_builder`]: crate::offers::flow::OffersMessageFlow::create_refund_builder
 	pub fn new(
 		metadata: Vec<u8>, signing_pubkey: PublicKey, amount_msats: u64
 	) -> Result<Self, Bolt12SemanticError> {
